+++ src/components/Modal/GroupMemberListModal.tsx
@@ -14,7 +14,6 @@ interface GroupMemberListProps {
 
 const GroupMemberListModal = ({ isOpen, handleModal, groupId }: GroupMemberListProps) => {
   const numGroupId = Number(groupId);
-
   const { data, isLoading } = useQuery({
     queryKey: GROUP_KEYS.members({ groupId: numGroupId }),
     queryFn: () => getGroupMemberFn(numGroupId),
